refactor(shared): US-ADJ-2.6 — Disciplina + DisciplinaDescriptor + UnidadMedida a shared/domain/
- Crea shared/domain/value_objects/{disciplina,unidad_medida,disciplina_descriptor}.py
  como fuentes canónicas (tipos cross-BC del dominio del deporte)
- competencia/domain/value_objects re-exporta desde shared (backward compat)
- resultados/domain/ y resultados/application/ importan desde shared.domain
- Crea resultados/infrastructure/repositories/disciplina_descriptor_adapter.py (fix B-02)
- Elimina violación cross-BC: resultados ya no importa de competencia.domain ni
  competencia.infrastructure para estos tipos
- 481 tests pasando, 0 regresiones
