+++ src/competencia/domain/value_objects/disciplina_descriptor.py
@@ -1,49 +1,4 @@
-"""Value Object DisciplinaDescriptor — reglas de medición y ordenamiento por disciplina."""
-from __future__ import annotations
+"""Re-export de DisciplinaDescriptor desde shared.domain (fuente canónica)."""
+from shared.domain.value_objects.disciplina_descriptor import DisciplinaDescriptor
 
-from dataclasses import dataclass
-
-from competencia.domain.value_objects.disciplina import Disciplina
-from competencia.domain.value_objects.unidad_medida import UnidadMedida
-
-
-@dataclass(frozen=True)
-class DisciplinaDescriptor:
-    """Encapsula las reglas de una disciplina: unidad esperada y orden de grilla.
-
-    Política P-01:
-        - STA (tiempo): unidad=Segundos, orden_ascendente=False (mayor AP primero)
-        - Distancia (DNF, DYN, ...): unidad=Metros, orden_ascendente=True (menor AP primero)
-
-    Attributes:
-        disciplina: Disciplina a la que pertenece este descriptor.
-        unidad_esperada: Unidad de medida válida para APs y RPs de esta disciplina.
-        orden_ascendente: True si la grilla ordena de menor a mayor AP (distancia);
-            False si ordena de mayor a menor (tiempo).
-    """
-
-    disciplina: Disciplina
-    unidad_esperada: UnidadMedida
-    orden_ascendente: bool
-
-    @classmethod
-    def para(cls, disciplina: Disciplina) -> DisciplinaDescriptor:
-        """Construye el descriptor canónico para una disciplina.
-
-        Args:
-            disciplina: Disciplina para la que se construye el descriptor.
-
-        Returns:
-            DisciplinaDescriptor con unidad y orden correctos según política P-01.
-        """
-        if disciplina.es_tiempo():
-            return cls(
-                disciplina=disciplina,
-                unidad_esperada=UnidadMedida.Segundos,
-                orden_ascendente=False,
-            )
-        return cls(
-            disciplina=disciplina,
-            unidad_esperada=UnidadMedida.Metros,
-            orden_ascendente=True,
-        )
+__all__ = ["DisciplinaDescriptor"]
