+++ src/resultados/api/router.py
@@ -8,7 +8,7 @@ from uuid import UUID
 from fastapi import APIRouter, Depends
 from fastapi.responses import JSONResponse
 
-from competencia.domain.value_objects.disciplina import Disciplina
+from shared.domain.value_objects.disciplina import Disciplina
 from competencia.infrastructure.event_store.sqlite_event_store import SQLiteEventStore
 from resultados.application.queries.obtener_ranking import (
     ObtenerRankingHandler,
